move `isBotUserAgent` to browser-utils and guard interactionsIntegration setup
