+++ packages/browser/src/index.bundle.tracing.replay.feedback.ts
@@ -27,10 +27,10 @@ export {
 
 export {
   browserTracingIntegration,
-  isBotUserAgent,
   startBrowserTracingNavigationSpan,
   startBrowserTracingPageLoadSpan,
 } from './tracing/browserTracingIntegration';
+export { isBotUserAgent } from '@sentry/browser-utils';
 export { elementTimingIntegrationShim as elementTimingIntegration };
 export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
