+++ packages/browser/src/index.bundle.tracing.replay.logs.metrics.ts
@@ -21,10 +21,10 @@ export {
 
 export {
   browserTracingIntegration,
-  isBotUserAgent,
   startBrowserTracingNavigationSpan,
   startBrowserTracingPageLoadSpan,
 } from './tracing/browserTracingIntegration';
+export { isBotUserAgent } from '@sentry/browser-utils';
 export { elementTimingIntegration } from '@sentry/browser-utils';
 export { reportPageLoaded } from './tracing/reportPageLoaded';
 export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
