+++ packages/browser/src/index.bundle.tracing.ts
@@ -28,10 +28,10 @@ export {
 
 export {
   browserTracingIntegration,
-  isBotUserAgent,
   startBrowserTracingNavigationSpan,
   startBrowserTracingPageLoadSpan,
 } from './tracing/browserTracingIntegration';
+export { isBotUserAgent } from '@sentry/browser-utils';
 export { elementTimingIntegrationShim as elementTimingIntegration };
 export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
