+++ packages/browser/src/index.ts
@@ -29,11 +29,10 @@ export { getFeedback, sendFeedback } from '@sentry/feedback';
 export { defaultRequestInstrumentationOptions, instrumentOutgoingRequests } from './tracing/request';
 export {
   browserTracingIntegration,
-  isBotUserAgent,
   startBrowserTracingNavigationSpan,
   startBrowserTracingPageLoadSpan,
 } from './tracing/browserTracingIntegration';
-export { elementTimingIntegration, getAbsoluteUrl } from '@sentry/browser-utils';
+export { elementTimingIntegration, getAbsoluteUrl, isBotUserAgent } from '@sentry/browser-utils';
 export { normalizeStringifyValue } from './normalizeStringifyValue';
 export { reportPageLoaded } from './tracing/reportPageLoaded';
 export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
